Sync ONNX OneHot

+++ onnx/backend/test/case/node/lrn.py
@@ -1,13 +1,15 @@
+# Copyright (c) ONNX Project Contributors
+#
 # SPDX-License-Identifier: Apache-2.0
+from __future__ import annotations
 
 import math
 
 import numpy as np
 
 import onnx
-
-from ..base import Base
-from . import expect
+from onnx.backend.test.case.base import Base
+from onnx.backend.test.case.node import expect
 
 
 class LRN(Base):
@@ -32,8 +34,8 @@ class LRN(Base):
             square_sum[n, c, h, w] = sum(
                 x[
                     n,
-                    max(0, c - int(math.floor((nsize - 1) / 2))) : min(
-                        5, c + int(math.ceil((nsize - 1) / 2)) + 1
+                    max(0, c - math.floor((nsize - 1) / 2)) : min(
+                        5, c + math.ceil((nsize - 1) / 2) + 1
                     ),
                     h,
                     w,
@@ -56,8 +58,8 @@ class LRN(Base):
             square_sum[n, c, h, w] = sum(
                 x[
                     n,
-                    max(0, c - int(math.floor((nsize - 1) / 2))) : min(
-                        5, c + int(math.ceil((nsize - 1) / 2)) + 1
+                    max(0, c - math.floor((nsize - 1) / 2)) : min(
+                        5, c + math.ceil((nsize - 1) / 2) + 1
                     ),
                     h,
                     w,
