+++ onnx/reference/ops/op_celu.py
@@ -1,5 +1,7 @@
+# Copyright (c) ONNX Project Contributors
+
 # SPDX-License-Identifier: Apache-2.0
-# pylint: disable=W0221
+from __future__ import annotations
 
 import numpy as np
 
@@ -14,4 +16,4 @@ def _vcelu1(x: np.ndarray, alpha: float = 1.0) -> np.ndarray:
 
 class Celu(OpRun):
     def _run(self, x, alpha=None):  # type: ignore
-        return (_vcelu1(x, alpha),)
+        return (_vcelu1(x, alpha).astype(x.dtype),)
