+++ onnx/reference/ops/op_gru.py
@@ -1,5 +1,7 @@
+# Copyright (c) ONNX Project Contributors
+
 # SPDX-License-Identifier: Apache-2.0
-# pylint: disable=R0913,R0914,W0221,W0613
+from __future__ import annotations
 
 import numpy as np
 
@@ -36,7 +38,7 @@ class CommonGRU(OpRun):
         H_t = H_0
         for x in np.split(X, X.shape[0], axis=0):
             gates = np.dot(x, gates_w) + np.dot(H_t, gates_r) + gates_b
-            z, r = np.split(gates, 2, -1)  # pylint: disable=W0632
+            z, r = np.split(gates, 2, -1)
             z = self.f(z)
             r = self.f(r)
             h_default = self.g(
@@ -75,14 +77,14 @@ class CommonGRU(OpRun):
         B=None,
         sequence_lens=None,
         initial_h=None,
-        activation_alpha=None,
-        activation_beta=None,
-        activations=None,
-        clip=None,
-        direction=None,
+        activation_alpha=None,  # noqa: ARG002
+        activation_beta=None,  # noqa: ARG002
+        activations=None,  # noqa: ARG002
+        clip=None,  # noqa: ARG002
+        direction=None,  # noqa: ARG002
         hidden_size=None,
         layout=None,
-        linear_before_reset=None,
+        linear_before_reset=None,  # noqa: ARG002
     ):
         # TODO: support overridden attributes.
         num_directions = W.shape[0]
@@ -121,7 +123,8 @@ class CommonGRU(OpRun):
             )
 
         Y, Y_h = self._step(X, R, B, W, H_0, num_directions=num_directions)
-        return (Y,) if self.n_outputs == 1 else (Y, Y_h)
+        Y = Y.astype(X.dtype)
+        return (Y,) if self.n_outputs == 1 else (Y, Y_h.astype(X.dtype))
 
 
 class GRU(CommonGRU):
