+++ onnx/reference/ops/op_negative_log_likelihood_loss.py
@@ -1,12 +1,16 @@
+# Copyright (c) ONNX Project Contributors
+
 # SPDX-License-Identifier: Apache-2.0
-# pylint: disable=R0912,R0913,W0221
+from __future__ import annotations
 
 import numpy as np
 
 from onnx.reference.op_run import OpRun
 
 
-def _compute_negative_log_likelihood_loss(x, target, weight=None, reduction="mean", ignore_index=None):  # type: ignore
+def _compute_negative_log_likelihood_loss(
+    x, target, weight=None, reduction="mean", ignore_index=None
+):  # type: ignore
     input_shape = x.shape
     if len(input_shape) == 1:
         raise RuntimeError(f"Unsupported shape {input_shape!r}.")
@@ -66,7 +70,7 @@ def _compute_negative_log_likelihood_loss(x, target, weight=None, reduction="mea
         loss = np.mean(loss)
     elif reduction == "sum":
         loss = np.sum(loss)
-    return (loss,)
+    return (loss.astype(x.dtype),)
 
 
 class NegativeLogLikelihoodLoss(OpRun):
