+++ onnx/reference/ops/op_non_zero.py
@@ -1,5 +1,7 @@
+# Copyright (c) ONNX Project Contributors
+
 # SPDX-License-Identifier: Apache-2.0
-# pylint: disable=W0221
+from __future__ import annotations
 
 import numpy as np
 
@@ -8,5 +10,6 @@ from onnx.reference.op_run import OpRun
 
 class NonZero(OpRun):
     def _run(self, x):  # type: ignore
-        res = np.vstack(np.nonzero(x))
+        # Specify np.int64 for Windows x86 machines
+        res = np.vstack(np.nonzero(x)).astype(np.int64)
         return (res,)
