+++ onnx/reference/ops/op_trilu.py
@@ -1,5 +1,7 @@
+# Copyright (c) ONNX Project Contributors
+
 # SPDX-License-Identifier: Apache-2.0
-# pylint: disable=W0221
+from __future__ import annotations
 
 import numpy as np
 
@@ -11,4 +13,4 @@ class Trilu(OpRun):
         k = 0 if k is None else int(k)
         if upper:  # type: ignore
             return (np.triu(x, k),)
-        return (np.tril(x, k),)
+        return (np.tril(x, k).astype(x.dtype),)
